doc, tools: extract getId
