+++ tools/doc/html.js
@@ -19,6 +19,7 @@ const linkJsTypeDocs = require('./lib/linkJsTypeDocs')
 const parseAPIHeader = require('./lib/parseAPIHeader')
 const getSection = require('./lib/getSection')
 const buildToc = require('./lib/buildToc')
+const getId = require('./lib/getId')
 
 module.exports = toHTML;
 
@@ -30,18 +31,3 @@ renderer.heading = function(text, level) {
 marked.setOptions({
   renderer: renderer
 });
-
-
-var idCounters = {};
-function getId(text) {
-  text = text.toLowerCase();
-  text = text.replace(/[^a-z0-9]+/g, '_');
-  text = text.replace(/^_+|_+$/, '');
-  text = text.replace(/^([^a-z])/, '_$1');
-  if (idCounters.hasOwnProperty(text)) {
-    text += '_' + (++idCounters[text]);
-  } else {
-    idCounters[text] = 0;
-  }
-  return text;
-}
