+++ tools/doc/lib/getId.js
@@ -1,0 +1,13 @@
+var idCounters = {};
+module.exports = function getId(text) {
+  text = text.toLowerCase();
+  text = text.replace(/[^a-z0-9]+/g, '_');
+  text = text.replace(/^_+|_+$/, '');
+  text = text.replace(/^([^a-z])/, '_$1');
+  if (idCounters.hasOwnProperty(text)) {
+    text += '_' + (++idCounters[text]);
+  } else {
+    idCounters[text] = 0;
+  }
+  return text;
+}
